firmware/requirements.txt
flask==2.3.2
flask-login==0.6.2
sensirion-i2c-sen5x
sensirion-i2c-driver
pillow
gpiozero
smbus2
